[FEAT #50 ] 최고가 ZSetOperations

+++ auction-service/src/main/java/com/farmted/auctionservice/domain/Bidding.java
@@ -17,8 +17,8 @@ public class Bidding extends TimeStamp {
     private Long biddingId;
     private String biddingUuid;
 
-    private Integer biddingPrice; // 일반 입찰
-    private Integer biddingAutoPrice; // 자동 입찰
+    private double biddingPrice; // 일반 입찰
+    private double biddingAutoPrice; // 자동 입찰
     private LocalDateTime biddingTime; // 응찰 시간
 
     private String memberUuid; // 응찰자
