+++ auction-service/src/main/java/com/farmted/auctionservice/dto/ResponseBiddingDto/BiddingResponseDto.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 
 @Getter @AllArgsConstructor @NoArgsConstructor
 public class BiddingResponseDto {
-    private Integer biddingPrice; // 일반 입찰 금액
+    private double biddingPrice; // 일반 입찰 금액
     private LocalDateTime biddingTime; // 입찰 시간
     private String memberUuid; // 응찰자
